pyproject.toml
[project]
name = "pyloto-corp"
version = "0.1.0"
description = "Atendimento inicial institucional/comercial da Pyloto via WhatsApp"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "python-json-logger>=2.0",
  "httpx>=0.25",
  "google-cloud-firestore>=2.16",
  "google-cloud-storage>=2.14",
  "google-cloud-secret-manager>=2.16",
  "google-cloud-tasks>=2.16",
  "cryptography>=42.0",
  "openai>=1.45",
  "redis>=5.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.4",
  "mypy>=1.9",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = ["B008"]

[tool.ruff.lint.per-file-ignores]
# SIM117: nested with intencionais para clareza de escopo em testes com múltiplos mocks
# Justificativa: combinar 5+ context managers em uma linha reduz legibilidade e dificulta debug
# Escopo restrito aos arquivos que realmente precisam (não global em tests/**)
"tests/test_llm_pipeline_e2e.py" = ["SIM117"]
"tests/integration/test_firestore_conversations.py" = ["SIM117"]

[tool.pytest.ini_options]
testpaths = ["tests"]
